renaming extractor.py to base.py and converter.py to exporter.py
