+++ src/datumaro/plugins/data_formats/coco/base.py
