+++ src/datumaro/plugins/data_formats/coco/exporter.py
