+++ src/datumaro/plugins/data_formats/cvat/base.py
