+++ src/datumaro/plugins/data_formats/icdar/base.py
