+++ src/datumaro/plugins/data_formats/icdar/exporter.py
