+++ src/datumaro/plugins/data_formats/mapillary_vistas/base.py
