+++ src/datumaro/plugins/data_formats/tf_detection_api/base.py
