+++ src/datumaro/plugins/data_formats/voc/base.py
