+++ tests/unit/test_cvat_format.py
@@ -20,8 +20,8 @@ from datumaro.components.dataset import Dataset
 from datumaro.components.environment import Environment
 from datumaro.components.extractor import DatasetItem
 from datumaro.components.media import Image
-from datumaro.plugins.data_formats.cvat.converter import CvatConverter
-from datumaro.plugins.data_formats.cvat.extractor import CvatImporter
+from datumaro.plugins.data_formats.cvat.base import CvatImporter
+from datumaro.plugins.data_formats.cvat.exporter import CvatConverter
 
 from tests.requirements import Requirements, mark_requirement
 from tests.utils.assets import get_test_asset_path
